# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zhalwadi Samaj API Documentation",
        "description": "API documentation for Zhalwadi Samaj mobile application",
        "contact": {
            "email": "support@zhalwadisamaj.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://parivar.test",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/announcement/get_all": {
            "get": {
                "tags": [
                    "Announcements"
                ],
                "summary": "Get All Announcements",
                "operationId": "a7ba7269130ef488dd74329ecf408140",
                "parameters": [
                    {
                        "name": "length",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/business/get_all": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get All Businesses",
                "operationId": "7103779284cc6dd066bc570571e13733",
                "parameters": [
                    {
                        "name": "length",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/donation/get_all": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get All Donations",
                "operationId": "949901c53003aa515882bb6c50b1877a",
                "parameters": [
                    {
                        "name": "length",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/gallery/get_all": {
            "get": {
                "tags": [
                    "Gallery"
                ],
                "summary": "Get All Galleries",
                "operationId": "529c35f610c5cc24724fb628a90e6598",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your bearer token in the format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}